[Refactor] createElem util 적용
